requirements.txt
Flask>=2.0.0
psycopg2>=2.0.0
python-dotenv>=0.21.0
rdkit>=2023.3.2

